agent_two/pyproject.toml
[project]
name = "agent-two"
version = "1.0.0"
description = "Multi-agent system with echo and reverse capabilities"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.1",
    "python-dotenv",
    "httpx",
    "fastapi",
    "acp-sdk",
] 